requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
RPi.GPIO==0.7.0